feat: Add react-autosuggest npm dependency and create Modal component
Add the "react-autosuggest" npm dependency to the project's package.json file. This will allow us to use the autosuggest feature in our application. Additionally, create a new Modal component with its associated styles. The Modal component will be used to display content in a fixed position overlay on the page. This change enhances the functionality and user experience of the application.